F705 - Lab_Sim/Trabajos/Ejemplo/requirements.txt
pygame==2.5.2
numpy==1.26.3
matplotlib==3.8.2
pyyaml==6.0.1
pytest==7.4.3
